Don't need my own HSV converter. Android has one.
