+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/opmode/ManualTestingOpMode.java
@@ -5,12 +5,17 @@ import org.firstinspires.ftc.twenty403.Hardware;
 import org.firstinspires.ftc.twenty403.Robot;
 import org.firstinspires.ftc.twenty403.command.autonomous.StartingPosition;
 
+import com.acmerobotics.dashboard.FtcDashboard;
+import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import com.technototes.library.structure.CommandOpMode;
 import com.technototes.library.util.Alliance;
 
-@TeleOp(name = "Meet TeleOp")
+@Disabled
+@TeleOp(name = "Manual Testing")
+@SuppressWarnings("unused")
 public class ManualTestingOpMode extends CommandOpMode {
     public Robot robot;
     public Controls controls;
@@ -18,6 +23,7 @@ public class ManualTestingOpMode extends CommandOpMode {
 
     @Override
     public void uponInit() {
+        telemetry = new MultipleTelemetry(telemetry, FtcDashboard.getInstance().getTelemetry());
         hardware = new Hardware(hardwareMap);
         robot = new Robot(hardware, Alliance.NONE, StartingPosition.NEUTRAL);
         controls = new Controls(driverGamepad, robot);
